+++ .flake8
@@ -1,7 +1,7 @@
 [flake8]
 ignore = E741,E226,W503,W504
 per-file-ignores =
-    */__init__.py:F401,F403
+    */__init__.py:F401,F403,F405
 exclude =
     build/,
     .eggs/,
