+++ pyccl/__init__.py
@@ -20,109 +20,32 @@ import numpy
 numpy.int = int if parse(numpy.__version__) >= parse("1.20.0") else numpy.int
 del parse, numpy
 
-# SWIG-generated
-from . import ccllib as lib
-
-# Name of the default power spectrum.
 DEFAULT_POWER_SPECTRUM = "delta_matter:delta_matter"
 
-# Hashing, Caching, CCL base, Mutation locks
+from . import ccllib as lib
+from .errors import *
 from .base import *
-
-# Errors
-from .errors import (
-    CCLError,
-    CCLWarning,
-    CCLDeprecationWarning,
-)
-
-# Constants and accuracy parameters
-from .parameters import *
-
-# Core data structures
+from .base.parameters import *
 from .cosmology import *
-
-# Background cosmology functions and growth functions
-from .background import (
-    growth_factor,
-    growth_factor_unnorm,
-    growth_rate,
-    comoving_radial_distance,
-    angular_diameter_distance,
-    comoving_angular_distance,
-    luminosity_distance,
-    distance_modulus,
-    h_over_h0,
-    scale_factor_of_chi,
-    omega_x,
-    rho_x,
-    sigma_critical,
-)
-
-# Boltzmann solvers
-from .boltzmann import (
-    get_camb_pk_lin,
-    get_isitgr_pk_lin,
-    get_class_pk_lin,
-)
-
-# Generalized power spectra
-from .pk2d import (
-    Pk2D,
-    parse_pk2d,
-)
-
-# Generalized connected trispectra
-from .tk3d import Tk3D
-
-# Power spectrum calculations, sigma8 and kNL
+from .background import *
+from .boltzmann import *
+from .pk2d import *
+from .tk3d import *
 from .power import *
-
-# Baryons & Neutrinos
-from .bcm import (
-    bcm_model_fka,
-    bcm_correct_pk2d,
-)
-
 from .neutrinos import *
-
-# Cells & Tracers
-from .cells import angular_cl
-from .tracers import (
-    Tracer,
-    NzTracer,
-    NumberCountsTracer,
-    WeakLensingTracer,
-    CMBLensingTracer,
-    tSZTracer,
-    CIBTracer,
-    ISWTracer,
-    get_density_kernel,
-    get_kappa_kernel,
-    get_lensing_kernel,
-)
-
-# Correlations & Covariances
-from .correlations import (
-    correlation,
-    correlation_3d,
-    correlation_multipole,
-    correlation_3dRsd,
-    correlation_3dRsd_avgmu,
-    correlation_pi_sigma,
-)
-
-from .covariances import (
-    angular_cl_cov_cNG,
-    angular_cl_cov_SSC,
-    sigma2_B_disc,
-    sigma2_B_from_mask,
-)
-
-# Miscellaneous
+from .cells import *
+from .tracers import *
+from .correlations import *
+from .covariances import *
 from .pyutils import debug_mode, resample_array
 
 # Deprecated & Renamed modules
+from .bcm import *
+from .halomodel import *
+from .massfunction import *
+from .haloprofile import *
+from .baryons import *
+
 def __getattr__(name):
     rename = {"core": "cosmology", "cls": "cells"}
     if name in rename:
@@ -133,55 +56,3 @@ def __getattr__(name):
         name = rename[name]
         return eval(name)
     raise AttributeError(f"No module named {name}.")
-
-from .halomodel import (
-    halomodel_matter_power,
-    halo_concentration,
-    onehalo_matter_power,
-    twohalo_matter_power,
-)
-
-from .massfunction import (
-    massfunc,
-    halo_bias,
-    massfunc_m2r,
-)
-
-from .haloprofile import (
-    nfw_profile_3d,
-    einasto_profile_3d,
-    hernquist_profile_3d,
-    nfw_profile_2d,
-)
-
-from .baryons import (
-    Baryons,
-    BaryonsSchneider15
-)
-
-
-__all__ = (
-    'lib',
-    'CCLError', 'CCLWarning', 'CCLDeprecationWarning',
-    'growth_factor', 'growth_factor_unnorm', 'growth_rate',
-    'comoving_radial_distance', 'angular_diameter_distance',
-    'comoving_angular_distance', 'luminosity_distance', 'distance_modulus',
-    'h_over_h0', 'scale_factor_of_chi', 'omega_x', 'rho_x', 'sigma_critical',
-    'get_camb_pk_lin', 'get_isitgr_pk_lin', 'get_class_pk_lin',
-    'Pk2D', 'parse_pk2d', 'Tk3D',
-    'bcm_model_fka', 'bcm_correct_pk2d',
-    'angular_cl',
-    'Tracer', 'NumberCountsTracer', 'WeakLensingTracer', 'CMBLensingTracer',
-    'tSZTracer', 'CIBTracer', 'ISWTracer', 'NzTracer',
-    'get_density_kernel', 'get_kappa_kernel', 'get_lensing_kernel',
-    'correlation', 'correlation_3d', 'correlation_multipole',
-    'correlation_3dRsd', 'correlation_3dRsd_avgmu', 'correlation_pi_sigma',
-    'angular_cl_cov_cNG', 'angular_cl_cov_SSC',
-    'sigma2_B_disc', 'sigma2_B_from_mask',
-    'debug_mode', 'resample_array',
-    'halomodel_matter_power', 'halo_concentration',
-    'onehalo_matter_power', 'twohalo_matter_power',
-    'massfunc', 'halo_bias', 'massfunc_m2r', 'nfw_profile_3d',
-    'einasto_profile_3d', 'hernquist_profile_3d', 'nfw_profile_2d',
-    'Baryons', 'BaryonsSchneider15',
-)
