+++ pyccl/baryons/__init__.py
@@ -1,8 +1,2 @@
-from .baryons_base import Baryons
-from .schneider15 import BaryonsSchneider15
-
-
-__all__ = (
-    "Baryons",
-    "BaryonsSchneider15"
-)
+from .baryons_base import *
+from .schneider15 import *
