+++ pyccl/base/__init__.py
@@ -1,3 +1,4 @@
+from .parameters import *
 from .caching import *
 from .schema import *
 from .deprecations import *
