+++ pyccl/halos/__init__.py
@@ -1,11 +1,11 @@
-from .halo_model_base import *
 from .concentration import *
-from .hmfunc import *
 from .hbias import *
-from .massdef import *
+from .hmfunc import *
 from .profiles import *
-from .profiles_2pt import *
+from .halo_model_base import *
 from .halo_model import *
+from .massdef import *
 from .pk_1pt import *
 from .pk_2pt import *
 from .pk_4pt import *
+from .profiles_2pt import *
