+++ pyccl/halos/halo_model.py
@@ -5,7 +5,7 @@ from ..pyutils import _spline_integrate
 from .. import background
 from ..base import (CCLAutoRepr, unlock_instance,
                     warn_api, deprecate_attr, deprecated)
-from ..parameters import physical_constants as const
+from ..base.parameters import physical_constants as const
 import numpy as np
