+++ pyccl/halos/profiles/__init__.py
@@ -1,9 +1,9 @@
 from .profile_base import *
 from .cib_shang12 import *
-from .gaussian import *
-from .powerlaw import *
-from .nfw import *
 from .einasto import *
+from .gaussian import *
 from .hernquist import *
-from .pressure_gnfw import *
 from .hod import *
+from .nfw import *
+from .powerlaw import *
+from .pressure_gnfw import *
