+++ pyccl/nl_pt/__init__.py
@@ -1,29 +1,3 @@
-# Tracers
-from .tracers import (
-    PTTracer,
-    PTMatterTracer,
-    PTNumberCountsTracer,
-    PTIntrinsicAlignmentTracer,
-    translate_IA_norm,
-)
-
-# Old power spectrum calculator
-# TODO v3: deprecate this
-from .power import (
-    PTCalculator,
-    get_pt_pk2d,
-)
-
-# Eulerian PT
-from .ept import EulerianPTCalculator
-
-__all__ = (
-    'PTTracer',
-    'PTMatterTracer',
-    'PTNumberCountsTracer',
-    'PTIntrinsicAlignmentTracer',
-    'translate_IA_norm',
-    'EulerianPTCalculator',
-    'PTCalculator',  # TODO v3: depr
-    'get_pt_pk2d',  # TODO v3: depr
-)
+from .ept import *
+from .power import *
+from .tracers import *
