+++ pyccl/nl_pt/tracers.py
@@ -1,9 +1,13 @@
 import numpy as np
 from scipy.interpolate import interp1d
 from ..pyutils import _check_array_params
-from ..parameters import physical_constants
+from ..base.parameters import physical_constants
 from ..base import CCLAutoRepr, warn_api
 from abc import abstractmethod
+
+
+__all__ = ("translate_IA_norm", "PTTracer", "PTMatterTracer",
+           "PTNumberCountsTracer", "PTIntrinsicAlignmentTracer",)
 
 
 @warn_api
