+++ pyccl/pyutils.py
@@ -1,8 +1,7 @@
 """Utility functions to analyze status and error messages passed from CCL, as
 well as wrappers to automatically vectorize functions."""
 from . import ccllib as lib
-from ._types import error_types
-from .parameters import spline_params
+from .base.parameters import spline_params
 from .errors import CCLError
 import numpy as np
 from collections.abc import Iterable
@@ -18,6 +17,25 @@ extrap_types = {'none': lib.f1d_extrap_0,
                 'linx_logy': lib.f1d_extrap_linx_logy,
                 'logx_liny': lib.f1d_extrap_logx_liny,
                 'logx_logy': lib.f1d_extrap_logx_logy}
+
+
+CLevelErrors = {
+    lib.CCL_ERROR_CLASS: 'CCL_ERROR_CLASS',
+    lib.CCL_ERROR_INCONSISTENT: 'CCL_ERROR_INCONSISTENT',
+    lib.CCL_ERROR_INTEG: 'CCL_ERROR_INTEG',
+    lib.CCL_ERROR_LINSPACE: 'CCL_ERROR_LINSPACE',
+    lib.CCL_ERROR_MEMORY: 'CCL_ERROR_MEMORY',
+    lib.CCL_ERROR_ROOT: 'CCL_ERROR_ROOT',
+    lib.CCL_ERROR_SPLINE: 'CCL_ERROR_SPLINE',
+    lib.CCL_ERROR_SPLINE_EV: 'CCL_ERROR_SPLINE_EV',
+    lib.CCL_ERROR_COMPUTECHI: 'CCL_ERROR_COMPUTECHI',
+    lib.CCL_ERROR_MF: 'CCL_ERROR_MF',
+    lib.CCL_ERROR_HMF_INTERP: 'CCL_ERROR_HMF_INTERP',
+    lib.CCL_ERROR_PARAMETERS: 'CCL_ERROR_PARAMETERS',
+    lib.CCL_ERROR_NU_INT: 'CCL_ERROR_NU_INT',
+    lib.CCL_ERROR_EMULATOR_BOUND: 'CCL_ERROR_EMULATOR_BOUND',
+    lib.CCL_ERROR_MISSING_CONFIG_FILE: 'CCL_ERROR_MISSING_CONFIG_FILE',
+}
 
 
 def check(status, cosmo=None):
@@ -40,8 +58,8 @@ def check(status, cosmo=None):
         msg = ""
 
     # Check for known error status
-    if status in error_types.keys():
-        raise CCLError("Error %s: %s" % (error_types[status], msg))
+    if status in CLevelErrors.keys():
+        raise CCLError("Error %s: %s" % (CLevelErrors[status], msg))
 
     # Check for unknown error
     if status != 0:
