+++ pyccl/tracers.py
@@ -5,10 +5,15 @@ import numpy as np
 from . import ccllib as lib
 from .pyutils import check
 from .errors import CCLWarning
-from .parameters import physical_constants
+from .base.parameters import physical_constants
 from .base import CCLObject, UnlockInstance, unlock_instance, warn_api
 from .pyutils import (_check_array_params, NoneArr, _vectorize_fn6,
                       _get_spline1d_arrays, _get_spline2d_arrays)
+
+
+__all__ = ("get_density_kernel", "get_lensing_kernel", "get_kappa_kernel",
+           "Tracer", "NzTracer", "NumberCountsTracer", "WeakLensingTracer",
+           "CMBLensingTracer", "tSZTracer", "CIBTracer", "ISWTracer",)
 
 
 def _Sig_MG(cosmo, a, k):
